fastapi_ai_chatbot/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1
openai==1.58.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.0
sse-starlette==2.1.3
jinja2==3.1.5
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.2.0
aioredis==2.0.1
markdown==3.7
bleach==6.2.0
greenlet==3.2.4